chore: rename folder/file to match export
