+++ src/pages/Maps/Content/MapView/Sprites.tsx
