+++ src/pages/Maps/Content/index.ts
@@ -1,4 +1,4 @@
-import { MapView } from './View'
+import { MapView } from './MapView/MapView'
 import { Controls } from './Controls/Controls'
 
 export { MapView, Controls }
